Refactor ODESolver/ODEVector
Split into header/source.
